app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "")
EMBEDDING_MODEL = os.getenv("EMBEDDING_MODEL", "text-embedding-3-small")
DATA_DIR = os.getenv("DATA_DIR", "data")
LOG_LEVEL = os.getenv("LOG_LEVEL", "INFO").upper()
