requirements.txt
openai
python-dotenv
pydantic
pymupdf
faiss-cpu
numpy
tqdm
